moved getFilteredSessions from SessionCtrl to SessionService
