+++ controllers/SessionCtrl.js
@@ -174,326 +174,6 @@ module.exports = function(socketService) {
     ) {
       const session = await Session.findById(sessionId)
       this.verifySessionParticipant(session, user, error)
-    },
-
-    getFilteredSessions: async function({
-      showBannedUsers,
-      showTestUsers,
-      minSessionLength,
-      sessionActivityFrom,
-      sessionActivityTo,
-      minMessagesSent,
-      studentRating,
-      volunteerRating,
-      firstTimeStudent,
-      firstTimeVolunteer,
-      isReported,
-      page
-    }) {
-      const PER_PAGE = 15
-      const pageNum = parseInt(page) || 1
-      const skip = (pageNum - 1) * PER_PAGE
-      const oneDayInMS = 1000 * 60 * 60 * 24
-      const estTimeOffset = 1000 * 60 * 60 * 4
-
-      // Add a day to the sessionActivityTo to make it inclusive for the activity range: [sessionActivityFrom, sessionActivityTo]
-      const inclusiveSessionActivityTo =
-        new Date(sessionActivityTo).getTime() + oneDayInMS + estTimeOffset
-      const offsetSessionActivityFrom =
-        new Date(sessionActivityFrom).getTime() + estTimeOffset
-
-      const sessionQueryFilter = {
-        // Filter by the length of a session
-        sessionLength: { $gte: parseInt(minSessionLength) * 60000 }
-      }
-      if (isReported) sessionQueryFilter.isReported = true
-
-      const ratingQueryFilter = {}
-      if (Number(studentRating))
-        ratingQueryFilter.studentRating = Number(studentRating)
-      if (Number(volunteerRating))
-        ratingQueryFilter.volunteerRating = Number(volunteerRating)
-
-      const userQueryFilter = {
-        'student.isTestUser': showTestUsers ? { $in: [true, false] } : false
-      }
-      if (firstTimeStudent && firstTimeVolunteer) {
-        userQueryFilter.$or = [
-          { 'student.totalPastSessions': 1 },
-          { 'volunteer.totalPastSessions': 1 }
-        ]
-      } else if (firstTimeStudent) {
-        userQueryFilter['student.totalPastSessions'] = 1
-      } else if (firstTimeVolunteer) {
-        userQueryFilter['volunteer.totalPastSessions'] = 1
-      }
-
-      try {
-        const sessions = await Session.aggregate([
-          {
-            $sort: {
-              createdAt: -1
-            }
-          },
-          {
-            $match: {
-              // Filter by a specific date range the sessions took place
-              createdAt: {
-                $gte: new Date(offsetSessionActivityFrom),
-                $lte: new Date(inclusiveSessionActivityTo)
-              },
-              // Filter a session by the amount of messages sent
-              $expr: {
-                $gte: [{ $size: '$messages' }, parseInt(minMessagesSent)]
-              }
-            }
-          },
-          {
-            $project: {
-              createdAt: 1,
-              endedAt: 1,
-              volunteer: { $ifNull: ['$volunteer', null] },
-              totalMessages: { $size: '$messages' },
-              type: 1,
-              subTopic: 1,
-              student: 1,
-              isReported: 1
-            }
-          },
-          {
-            $addFields: {
-              // Add the length of a session on the session documents
-              sessionLength: {
-                $cond: {
-                  if: { $ifNull: ['$endedAt', undefined] },
-                  then: { $subtract: ['$endedAt', '$createdAt'] },
-                  // $$NOW is a mongodb system variable which returns the current time
-                  else: { $subtract: ['$$NOW', '$createdAt'] }
-                }
-              },
-              volunteer: {
-                $cond: {
-                  if: { $ifNull: ['$volunteer', undefined] },
-                  then: '$volunteer',
-                  else: null
-                }
-              }
-            }
-          },
-          {
-            $match: sessionQueryFilter
-          },
-          {
-            $lookup: {
-              from: 'feedbacks',
-              localField: '_id',
-              foreignField: 'sessionId',
-              as: 'feedbacks'
-            }
-          },
-          // add student and volunteer feedback if present
-          {
-            $addFields: {
-              studentFeedback: {
-                $filter: {
-                  input: '$feedbacks',
-                  as: 'feedback',
-                  cond: { $eq: ['$$feedback.userType', 'student'] }
-                }
-              },
-              volunteerFeedback: {
-                $filter: {
-                  input: '$feedbacks',
-                  as: 'feedback',
-                  cond: { $eq: ['$$feedback.userType', 'volunteer'] }
-                }
-              }
-            }
-          },
-          {
-            $unwind: {
-              path: '$studentFeedback',
-              preserveNullAndEmptyArrays: true
-            }
-          },
-          {
-            $unwind: {
-              path: '$volunteerFeedback',
-              preserveNullAndEmptyArrays: true
-            }
-          },
-          {
-            $addFields: {
-              studentRating: {
-                $cond: {
-                  if: '$studentFeedback.responseData.rate-session.rating',
-                  then: '$studentFeedback.responseData.rate-session.rating',
-                  else: null
-                }
-              },
-              volunteerRating: {
-                $cond: {
-                  if: '$volunteerFeedback.responseData.rate-session.rating',
-                  then: '$volunteerFeedback.responseData.rate-session.rating',
-                  else: null
-                }
-              }
-            }
-          },
-          {
-            $match: ratingQueryFilter
-          },
-          {
-            $lookup: {
-              from: 'users',
-              let: {
-                studentId: '$student'
-              },
-              pipeline: [
-                {
-                  $match: {
-                    $expr: {
-                      $eq: ['$_id', '$$studentId']
-                    }
-                  }
-                },
-                {
-                  $project: {
-                    firstname: 1,
-                    isBanned: 1,
-                    isTestUser: 1,
-                    totalPastSessions: { $size: '$pastSessions' }
-                  }
-                }
-              ],
-              as: 'student'
-            }
-          },
-          {
-            $unwind: '$student'
-          },
-          {
-            $lookup: {
-              from: 'users',
-              let: {
-                volunteerId: '$volunteer'
-              },
-              pipeline: [
-                {
-                  $match: {
-                    $expr: {
-                      $eq: ['$_id', '$$volunteerId']
-                    }
-                  }
-                },
-                {
-                  $project: {
-                    firstname: 1,
-                    totalPastSessions: { $size: '$pastSessions' }
-                  }
-                }
-              ],
-              as: 'volunteer'
-            }
-          },
-          {
-            $unwind: {
-              path: '$volunteer',
-              preserveNullAndEmptyArrays: true
-            }
-          },
-          {
-            $match: userQueryFilter
-          },
-          {
-            $lookup: {
-              from: 'useractions',
-              let: { userId: '$student._id' },
-              pipeline: [
-                // Get user actions with 'BANNED' per session
-                {
-                  $match: {
-                    $expr: {
-                      $and: [
-                        { $eq: ['$user', '$$userId'] },
-                        { $eq: ['$action', USER_ACTION.ACCOUNT.BANNED] }
-                      ]
-                    }
-                  }
-                }
-              ],
-              as: 'bannedUserAction'
-            }
-          },
-          {
-            $addFields: {
-              // Retrieve the most recent 'BANNED' user action
-              lastBannedAt: { $max: '$bannedUserAction.createdAt' }
-            }
-          },
-          {
-            // Show sessions that were created before a user has been banned
-            // If showBannedUsers is true, show all sessions up to chosen date
-            $addFields: {
-              showSession: {
-                $cond: {
-                  if: {
-                    $or: [
-                      { $eq: ['$lastBannedAt', undefined] },
-                      { $eq: ['$student.isBanned', false] }
-                    ]
-                  },
-                  then: true,
-                  else: {
-                    $cond: [
-                      {
-                        $lte: [
-                          '$createdAtEstTime',
-                          showBannedUsers
-                            ? new Date(inclusiveSessionActivityTo)
-                            : '$lastBannedAt'
-                        ]
-                      },
-                      true,
-                      false
-                    ]
-                  }
-                }
-              }
-            }
-          },
-          {
-            $match: {
-              showSession: true
-            }
-          },
-          {
-            $skip: skip
-          },
-          {
-            $limit: PER_PAGE
-          },
-          {
-            $project: {
-              createdAt: 1,
-              endedAt: 1,
-              volunteer: 1,
-              totalMessages: 1,
-              type: 1,
-              subTopic: 1,
-              student: 1,
-              studentFirstName: '$student.firstname',
-              studentRating: 1
-            }
-          }
-        ])
-
-        const isLastPage = sessions.length < PER_PAGE
-
-        return { sessions, isLastPage }
-      } catch (err) {
-        throw new Error(err.message)
-      }
     }
   }
 }
